Directory structure Organized the new FlatBuffers backend. First, I separated Image.fbs out into its own file. I renamed the new backend folder to "ipc" per Josh's request. I also moved the codegen to a subfolder called "codegen".
