+++ CompilerSource/ipc/Background.fbs
@@ -1,0 +1,22 @@
+include "Image.fbs";
+
+table Background {
+  name:string;
+  id:int;
+
+  transparent:bool;
+  smooth_edges:bool;
+  preload:bool;
+  use_as_tileset:bool;
+
+  tile_width:int;
+  tile_height:int;
+  horizontal_offset:int;
+  vertical_offset:int;
+  horizontal_separation:int;
+  vertical_separation:int;
+
+  image:Image;
+}
+
+root_type Background;
